Create averagejoinfreshness.java
Index construction:
We need 3 indexes first one to store individual predicate freshness[P]. The other 2 is used to store average freshness.  
1)	To compute average freshness over Subject we build a hashmap index [SP] with subject as key and corresponding value is a hashmap of all predicates that appeared with key along with their freshness [insertToSP function].  
2)	To compute average freshness over Object we build a hashmap index [OP] with object as key and corresponding value is a hashmap of all predicates that appeared with key along with their freshness [insertToOP function].  
In [functions AVG**UsingIndex] for 2 triple patterns of each join we compute both real freshness of that join and average freshness over triple pattern say (?s #p ?o), that is we iterate through subjects in SP and see if that subject’s predicate list contains #p and if so we consider its freshness in average.